+++ plugins/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
-- Added product flavor dimension support to environment config plugin.
+- Added product flavor dimension support to environment config plugin. 
 
 ## [1.3.0]